Agent/tutor2/requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
openai>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pymysql>=1.1.0
pypdf>=4.0.0
langchain-community>=0.2.0
duckduckgo-search>=6.2.0
langsmith>=0.1.95
